Use GrammaticalTermMap instead of raw Map in LanguageDictionary.
This is to be able to add some more variation of LanguageDictionary implementation.
Currently, it always expects all terms are in-memory and not allowing any changes as it is final class.
Create an abstract class GrammaticalTermMap instead of raw Map to enable other implemenation,
where can be more memory-efficient.

To create a new implementation extends GrammaticalLabelSetLoader to use new hooks. in it.
It can override a new instance or finalized instance after label loading.